Replace the SlideshowVideo internal component with the powerfull FluentCxVideo

+++ src/Core/Components/Media/Audio/AudioControls.razor.cs
@@ -87,7 +87,7 @@ public partial class AudioControls : FluentComponentBase
     /// Gets or sets the callback to be invoked when the volume changes.
     /// </summary>
     [Parameter]
-    public EventCallback<double> OnVolumeChanged { get; set; }
+    public EventCallback<int> OnVolumeChanged { get; set; }
 
     /// <summary>
     /// Gets or sets the callback to be invoked when the properties button is clicked.
